pyproject.toml
[tool.poetry]
name = "fastapi-transcript"
version = "0.1.0"
description = ""
authors = ["Hasnain Hasib <hasibjoy332@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
pydantic = "^2.8.2"
requests = "^2.32.3"
gunicorn = "^23.0.0"
openai = "0.28.0"
jinja2 = "3.1.2"
python-dotenv = "0.20.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
